+++ sample/authui/src/custom.js
@@ -21,8 +21,8 @@ import 'bootstrap';
 // Import Firebase dependencies.
 import firebase from '@firebase/app';
 import '@firebase/auth';
-// Import GCIP/IAP module (using local build).
-import * as ciap from '../../../dist/index.esm';
+// Import GCIP/IAP module.
+import * as ciap from '../../../builds/ciap/index.esm';
 
 import * as templates from './templates';
 import jQuery from 'jquery';